backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
eventkit==1.0.3
fastapi==0.115.8
h11==0.14.0
ib_async==1.0.3
idna==3.10
nest-asyncio==1.6.0
numpy==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
